Player Service Full Implementations, More Refactorings and Forntend Implementations

+++ src/backend/league_service/src/main/java/ies/gamesense/league_service/repositories/LeagueRepository.java
@@ -1,14 +1,11 @@
 package ies.gamesense.league_service.repositories;
 import ies.gamesense.league_service.entities.League;
-import ies.gamesense.league_service.entities.League_Club;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface LeagueRepository extends JpaRepository<League, Integer> {
+public interface LeagueRepository extends JpaRepository<League, Long> {
     League findByName(String name);
     boolean existsByName(String name);    
 }
